study_bot/requirements.txt
aiogram==3.24.0
gigachat==0.2.0
python-dotenv==1.2.1